fareharbor-salty/requirements.txt
Django==1.11.29
django-extensions==3.1.1
djangorestframework==3.8.2
requests==2.20.0
